Fix missing addrview error
ZoneMTA wildduck plugin uses wildduck's user handler for address resolution.
A missing field in the projections map caused envelop from addresses
in outgoing messages that matched a wildcard address of the user to
be rejected by ZoneMTA wildduck plugin because of the error thrown in
the user handler of wildduck.